requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
babel==2.17.0
bcrypt==4.3.0
bleach==6.2.0
blinker==1.9.0
cffi==1.17.1
click==8.2.1
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.1
flask-babel==4.0.0
Flask-Login==0.6.3
Flask-Mailman==1.1.1
Flask-Principal==0.4.0
Flask-Security==5.6.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.3
idna==3.10
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
passlib==1.7.4
pycparser==2.22
python-dotenv==1.1.1
pytz==2025.2
SQLAlchemy==2.0.41
typing_extensions==4.14.0
webencodings==0.5.1
Werkzeug==3.1.3
WTForms==3.2.1
